binance_query/requirements.txt
asyncio
python-dotenv
fastapi
uvicorn
pydantic
httpx
cryptography